Feature/add game service (#11)
* feat: add map seeder, update graph service and tests

* feat: update event turn logic and tests

* feat: update delivery turn and tests

* feat: add game service and update models, services, and tests

* refactor: ruff fix

* refactor: fir requirements.txt

* feat: add GameService and integration tests

- Recreate unit and connector integration tests
- Add service integration tests
- Add mypy and Ruff format checks to CI
- Update requirements
- Fix service issues

* fix: reuirements

+++ backend/app/models/road.py
@@ -25,7 +25,7 @@ class Road(Base):
     )
 
     # Physical length of the road in kilometers.
-    distance: Mapped[float] = mapped_column(
+    distance: Mapped[int] = mapped_column(
         nullable=False,
     )
